+++ extension.js
@@ -6,7 +6,33 @@ const messages = {
   error: () => vscode.window.showErrorMessage('Error message!', { title: 'Yes' }, { title: 'No' })
 };
 
+const progressBadge = {
+  _resolve: null,
+  show() {
+    if (this._resolve) {
+      vscode.window.showWarningMessage('Progress Badge is already visible in Source Control!');
+      return;
+    }
+
+    vscode.window.withProgress({ location: vscode.ProgressLocation.SourceControl }, () => {
+      vscode.window.showInformationMessage('Progress Badge is visible in Source Control.');
+      return new Promise(resolve => (this._resolve = resolve));
+    });
+  },
+  stop() {
+    if (!this._resolve) {
+      vscode.window.showWarningMessage('Progress Badge is already stopped!');
+      return;
+    }
+
+    this._resolve();
+    this._resolve = null;
+    vscode.window.showInformationMessage('Progress Badge is stopped.');
+  }
+};
+
 function activate(context) {
+  // Notification Messages
   const showInformationMessage = vscode.commands.registerCommand('show.information.message', messages.information);
   const showWarningMessage = vscode.commands.registerCommand('show.warning.message', messages.warning);
   const showErrorMessage = vscode.commands.registerCommand('show.error.message', messages.error);
@@ -18,6 +44,13 @@ function activate(context) {
   context.subscriptions.push(showWarningMessage);
   context.subscriptions.push(showErrorMessage);
   context.subscriptions.push(showAllMessages);
+
+  // Progress Badge
+  const showProgressBadgeInSourceControl = vscode.commands.registerCommand('show.progress.badge', progressBadge.show);
+  const stopProgressBadgeInSourceControl = vscode.commands.registerCommand('stop.progress.badge', progressBadge.stop);
+
+  context.subscriptions.push(showProgressBadgeInSourceControl);
+  context.subscriptions.push(stopProgressBadgeInSourceControl);
 }
 
 exports.activate = activate;